refactor: Reorganize folder structure
