+++ pkg/common/interfaces.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"context"
