+++ pkg/flags/flags.go
@@ -1,4 +1,4 @@
-package cmd
+package flags
 
 import (
 	"github.com/spf13/cobra"
